gRPC/project/requirements.txt
grpcio==1.41.1
grpcio-tools==1.41.1
protobuf==3.18.0
six==1.16.0
